Fix problem that avgOut runs no longer worked
removed more classes by puting the common code into the subclass

+++ srcTools/magma/tools/benchmark/view/bench/kickchallenge/KickBenchmarkTableModelExtension.java
@@ -79,7 +79,7 @@ class KickBenchmarkTableModelExtension extends DefaultTableModel
 	{
 		super(data, columnNames);
 
-		columnTypes = new Class[] {String.class, Object.class, Double.class, Integer.class, Integer.class, Float.class,
+		columnTypes = new Class[] {String.class, Object.class, Double.class, Integer.class, Integer.class, Double.class,
 				String.class, Float.class};
 
 		editableColumns = new boolean[] {true, false, false, false, false, false, true, true};
